# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Wharf Backend API",
        "description": "Zero-knowledge sync backend for Wharf. The server only ever stores\nciphertext vault blobs and bcrypt hashes of client-derived credential\nkeys. It never sees passwords or plaintext vault contents.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:18080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/users/me/public-key": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Publish or rotate the account's X25519 public key",
                "description": "Rotating (rotate=true) replaces the key and resets every wrapped key the account holds \u2014 each affected project membership re-enters the awaiting-key state.",
                "operationId": "updatePublicKey",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePublicKeyRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Key is not 32 bytes"
                    },
                    "409": {
                        "description": "A key already exists and rotate was not set"
                    },
                    "204": {
                        "description": "Public key stored"
                    },
                    "401": {
                        "description": "Not authenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/users/me/invites/{id}/decline": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Decline a project invite",
                "operationId": "declineInvite",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Invite not found or not addressed to the caller"
                    },
                    "204": {
                        "description": "Invite declined"
                    },
                    "401": {
                        "description": "Not authenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/me/invites/{id}/accept": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Accept a project invite, joining as an awaiting-key member",
                "operationId": "acceptInvite",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Invite not found or not addressed to the caller",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectSummary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectSummary"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Joined; returns the project summary",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectSummary"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "Invite expired",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectSummary"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/users/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get the authenticated account's profile",
                "description": "Includes hasPassword / hasRecovery / hasVault flags and the account's public key.",
                "operationId": "getCurrentUser",
                "responses": {
                    "200": {
                        "description": "Profile returned",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfile"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfile"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/me/invites": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List unexpired project invites addressed to the caller",
                "operationId": "getMyInvites",
                "responses": {
                    "200": {
                        "description": "Invites returned",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MyInvite"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MyInvite"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdatePublicKeyRequest": {
                "required": [
                    "publicKey"
                ],
                "type": "object",
                "properties": {
                    "publicKey": {
                        "type": "string",
                        "description": "Base64-encoded X25519 public key (exactly 32 bytes)"
                    },
                    "rotate": {
                        "type": "boolean",
                        "description": "When true, replace an existing key and reset every wrapped key the account holds (each affected project membership re-enters the awaiting-key state)"
                    }
                },
                "description": "Publish or rotate the account's X25519 public key"
            },
            "ProjectSummary": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "description": "The calling member's role in this project",
                        "enum": [
                            "MEMBER",
                            "ADMIN",
                            "OWNER"
                        ]
                    },
                    "memberCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "pendingInviteCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "vaultVersion": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "awaitingKey": {
                        "type": "boolean",
                        "description": "True when the caller holds no wrapped key yet (awaiting an admin to seal the DEK)"
                    }
                },
                "description": "A project as it appears in the caller's project list"
            },
...
            "UserProfile": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "email": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "hasPassword": {
                        "type": "boolean",
                        "description": "Whether a master-password auth key is set (false for a fresh OAuth account)"
                    },
                    "hasRecovery": {
                        "type": "boolean",
                        "description": "Whether a recovery key is set"
                    },
                    "hasVault": {
                        "type": "boolean",
                        "description": "Whether an encrypted vault exists for this account"
                    },
                    "publicKey": {
                        "type": "string",
                        "description": "The account's base64 X25519 public key, or null if none is published yet"
                    }
                },
                "description": "The authenticated account's profile plus routing capability flags"
            },
            "MyInvite": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "projectId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "projectName": {
                        "type": "string"
                    },
                    "invitedByEmail": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "expiresAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "description": "An invite addressed to the calling account"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Identity (access) token issued by /api/v1/auth.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}